Project Import
After some trouble with Git and the Oculus framework, it should all be ready now. Contains test environment scene with ability to pick up and throw cubes, as well as targets that register if cubes enter their area and then update their corresponding text elements to provide a hit counter.